// src/app/models/location.ts
export class ILocation {
    asn: string;
    city: string;
    continent_code: string;
    country: string;
    country_calling_code: string;
    country_name: string;
    currency: string;
    in_eu: string; false
    ip: string;
    languages: string;
    latitude: string;
    longitude: string;
    org: string;
    postal: string;
    region: string;
    region_code: string;
    timezone: string;
    utc_offset: string;
}